Add location services file and dependencies #29
Added a service file in ./services which provides an async function for getting location information. Also modified the map component to include a small button to test the feature. the changes to this file won't be merged and are just there to make it easier to test the functionality. Finally, also added an empty test file.

I also took the liberty to add types and code completion functionality for Web Storm, and to rename one of our components to better follow our new conventions

Issue number:  #29